refactor(xdl): extract findP12CertSerialNumber and getP12CertFingerprint to PKCS12Utils
This allows these functions to be used from credentials/IosCredentials without converting
the whole IosSigning module to TypeScript right now.
